Fixed bug with several classes having save issues when fully deep copying

+++ Scripts/SaveHandler.gd
@@ -5,7 +5,7 @@ var save_file_location: String = "user://save/"
 var save_exists_file: String = save_file_location + "playerinfo.tres" # this file should always exist if a save game exists
 var battle_file = save_file_location + "battle.tres" # this file should exist if the game was saved in battle
 
-var subdirs: Array = ["npcs/", "enemies/", "./"]
+var subdirs: Array = ["npcs/", "enemies/", ""] # last one is root save folder
 
 # Called when the node enters the scene tree for the first time.
 func _ready():
@@ -49,13 +49,15 @@ func load_data():
 func new_game():
 	fetch_saved_scripts()
 	for subdir in subdirs: # for each save subdirectory
-		var files = DirAccess.get_files_at(save_file_location + subdir)
-		for filepath in files: # remove all files inside the subdir
-			DirAccess.remove_absolute(save_file_location + subdir + filepath)
-		var err = DirAccess.remove_absolute(save_file_location + subdir) # then remove the subdir itself
-		if err > 0:
-			print("Remove save subdir" + subdir + " error " + String.num(err))
+		if DirAccess.dir_exists_absolute(save_file_location + subdir):
+			var files = DirAccess.get_files_at(save_file_location + subdir)
+			for filepath in files: # remove all files inside the subdir
+				DirAccess.remove_absolute(save_file_location + subdir + filepath)
+			var err = DirAccess.remove_absolute(save_file_location + subdir) # then remove the subdir itself
+			if err > 0:
+				print("Remove save subdir ", subdir, " error ", err)
 	
+	create_save_subdirs()
 	for script_path in saved_scripts:
 		var scr = get_node_or_null(NodePath(script_path))
 		if scr != null and scr.has_method("new_game"):
@@ -68,8 +70,13 @@ func is_save_in_battle():
 	return FileAccess.file_exists(battle_file)
 
 func create_save_subdirs():
+	if not DirAccess.dir_exists_absolute(save_file_location):
+		var err = DirAccess.make_dir_absolute(save_file_location)
+		if err > 0:
+			print("DirAccess create root save dir error ", err)
+	
 	for subdir in subdirs:
 		if not DirAccess.dir_exists_absolute(save_file_location + subdir):
 			var err = DirAccess.make_dir_absolute(save_file_location + subdir)
 			if err > 0:
-				print("DirAccess create dir " + subdir + " error " + String.num(err))
+				print("DirAccess create dir ", subdir, " error ", err)
